extract module controllers (#15)
* rename controller into controllers

* fix tests

* fix doc
