+++ council/agents/agent.py
@@ -3,12 +3,11 @@ from typing import List
 
 from council.core.budget import Budget
 from council.core.chain import Chain
-from council.core.controller_base import ControllerBase
+from council.controllers import ControllerBase, BasicController
 from council.core.evaluator_base import EvaluatorBase
 from council.core.execution_context import AgentContext, ChatHistory
 from council.core.runners import new_runner_executor
 from .agent_result import AgentResult
-from ..controller import BasicController
 from ..core import SkillBase
 from ..evaluator import BasicEvaluator
