+++ council/controllers/__init__.py
@@ -1,4 +1,5 @@
 """This package contains ready to use controllers"""
 
+from .controller_base import ControllerBase
 from .basic_controller import BasicController
 from .llm_controller import LLMController
