--- docs/source/reference/controller.rst
@@ -1,14 +1,0 @@
-Controller
-==========
-
-.. automodule:: council.controller
-
-.. autoclasstree:: council.controller
-    :full:
-    :namespace: council
-
-.. toctree::
-    :maxdepth: 1
-    :glob:
-
-    controller/*
