--- docs/source/reference/controller/basic_controller.rst
@@ -1,8 +1,0 @@
-BasicController
----------------
-
-.. autoclasstree:: council.controller.BasicController
-    :full:
-    :namespace: council
-
-.. autoclass:: council.controller.BasicController
