--- docs/source/reference/controller/llm_controller.rst
@@ -1,8 +1,0 @@
-LLMController
--------------
-
-.. autoclasstree:: council.controller.LLMController
-    :full:
-    :namespace: council
-
-.. autoclass:: council.controller.LLMController
