+++ docs/source/reference/controllers.rst
@@ -1,0 +1,14 @@
+Controllers
+-----------
+
+.. automodule:: council.controllers
+
+.. autoclasstree:: council.controllers
+    :full:
+    :namespace: council
+
+.. toctree::
+    :maxdepth: 1
+    :glob:
+
+    controllers/*
