+++ docs/source/reference/controllers/basic_controller.rst
@@ -1,0 +1,8 @@
+BasicController
+---------------
+
+.. autoclasstree:: council.controllers.BasicController
+    :full:
+    :namespace: council
+
+.. autoclass:: council.controllers.BasicController
