+++ docs/source/reference/controllers/controller_base.rst
@@ -1,0 +1,4 @@
+ControllerBase
+--------------
+
+.. autoclass:: council.controllers.ControllerBase
