+++ docs/source/reference/controllers/llm_controller.rst
@@ -1,0 +1,8 @@
+LLMController
+-------------
+
+.. autoclasstree:: council.controllers.LLMController
+    :full:
+    :namespace: council
+
+.. autoclass:: council.controllers.LLMController
