--- docs/source/reference/core/controller_base.rst
@@ -1,4 +1,0 @@
-ControllerBase
---------------
-
-.. autoclass:: council.core.ControllerBase
