+++ tests/integration/agent_tests/test_agent_test_suite.py
@@ -5,7 +5,7 @@ import dotenv
 
 from council.agents import Agent
 from council.agent_tests import AgentTestSuite, AgentTestCase
-from council.controller import LLMController
+from council.controllers import LLMController
 from council.core import Chain
 from council.evaluator import LLMEvaluator
 from council.llm import AzureLLM, AzureConfiguration
