+++ tests/integration/prompt/test_prompt.py
@@ -3,7 +3,7 @@ import unittest
 import dotenv
 
 from council.agents import Agent
-from council.controller import BasicController
+from council.controllers import BasicController
 from council.core import Chain, ChatHistory, AgentContext, Budget
 from council.evaluator import BasicEvaluator
 from council.llm import AzureConfiguration, AzureLLM
