+++ tests/integration/skill/test_llm_skill.py
@@ -4,7 +4,7 @@ from typing import List
 import dotenv
 
 from council.agents import Agent
-from council.controller import BasicController
+from council.controllers import BasicController
 from council.core import Chain, ChatHistory, AgentContext, Budget
 from council.evaluator import BasicEvaluator
 from council.llm import AzureConfiguration, AzureLLM, LLMMessage
