+++ tests/unit/agents/test_agent.py
@@ -1,7 +1,7 @@
 from unittest import TestCase
 
 from council.agents import Agent
-from council.controller import BasicController
+from council.controllers import BasicController
 from council.evaluator import BasicEvaluator
 from council.mocks import MockSkill
