+++ tests/unit/controllers/test_llm_controller.py
@@ -1,7 +1,7 @@
 import unittest
 from typing import List, Any
 
-from council.controller import LLMController
+from council.controllers import LLMController
 from council.core import Chain, Budget
 from council.core.execution_context import AgentContext, ChatHistory
 from council.llm import LLMMessage, LLMBase
